+++ timer.h
@@ -1,4 +1,7 @@
-extern unsigned long g_seconds;
+#include <stdint.h>
+
+
+extern uint32_t g_seconds;
 
 void initTimer0();
 void initTimer2();
